Fix installation errors due to missing packages
This commit fixes installation errors due to the install_requirements
specified not being available during install time when setup.py is run.
To fix this a static .version file is created, removing the need for
having the package version to be specified in __init__.py of the
pyscopus package.

Summary of Changes:
- Removed __version__ attribute from __init__.py
- Added static .version file to contain the current package version
- Removed import of pyscopus in setup.py to fix installation error
- Changed `HERE` directory to `ROOT_DIR` and preferred use of Path
  class over os.path.abspath